requirements.txt
# YSense Platform v3.0 Requirements
# Production Dependencies - FIXED VERSION

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# Comment out PostgreSQL if using SQLite locally
# psycopg2-binary==2.9.9
# asyncpg==0.29.0

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
cryptography==41.0.7
python-dotenv==1.0.0

# API & HTTP
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduler for orchestrator
apscheduler==3.10.4

# JSON Web Tokens
PyJWT==2.8.0

# Optional - Comment out if not needed
# redis==5.0.1
# aioredis==2.0.1
# aiosmtplib==3.0.1
# email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# UI Framework
streamlit==1.29.0
streamlit-extras==0.3.6

# Remove these - they don't exist as packages
# QWEN==1.3.8  # This is not a real package
# anthropic==0.7.8  # Use 'anthropic' without version if needed

# For API integration (if using Anthropic)
# anthropic

# Development Dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1